Load avatar urls for speakers in the firebase layer
Fixes #26:

When loading the speaker list, avatar images would pop in late, or have
a "ghost" from the previous bind. The viewholder was requesting
a `downloadUrl` for each avatar on bind. This switches to using
coroutines to map the avatar download url into the `Speaker` model
itself. Doing this lets the recycler have all the data it needs to bind
its viewholders.