Move setting of pager adapter so that list fragment state is restored after config change
After a config change (such as screen rotation),
a ViewPager2's state is restored when its adapter is set.  If the adapter is set before
the backing data is set, then each of the ViewPager2's page Fragment's states won't
be restored as expected.  By moving the setting of the adapter after the adapter's
data is set, it allows the currently selected tab to be restored as well as the scroll
position of the RecyclerViews in the list Fragments.